+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+xxx
+==================
+- Updated DEFAULT_USER_AGENT and USER_AGENT_LIST
+
 0.6.4 / 2018-09-05
 ==================
 - Added `compileScript()` function to `utils.puppeteer` to enable use of external scripts at runtime.